[refactor] NovelQueryDto 명칭 변경 - #31
QueryDto -> QueryProjection
