+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/NovelQueryRepository.java
@@ -1,8 +1,8 @@
 package com.iucyh.novelservice.novel.repository.query;
 
 import com.iucyh.novelservice.novel.enumtype.NovelCategory;
-import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
-import com.iucyh.novelservice.novel.repository.query.condition.NovelSearchCondition;
+import com.iucyh.novelservice.novel.repository.query.projection.NovelQueryProjection;
+import com.iucyh.novelservice.novel.repository.query.condition.NovelPagingCondition;
 import com.iucyh.novelservice.novel.repository.query.paging.NovelPagingStrategy;
 
 import java.util.List;
@@ -12,11 +12,11 @@ public interface NovelQueryRepository {
     /**
      * @param category 필터링 할 카테고리, 모든 카테고리 조회 시 null 전달
      */
-    List<? extends NovelQueryDto> findNovels(NovelSearchCondition condition, NovelPagingStrategy strategy, NovelCategory category);
+    List<? extends NovelQueryProjection> findNovels(NovelPagingCondition condition, NovelPagingStrategy strategy, NovelCategory category);
 
     /**
      * 이번달 신작 소설 조회 메서드
      * @param category 필터링 할 카테고리, 모든 카테고리 조회 시 null 전달
      */
-    List<? extends NovelQueryDto> findNewNovels(NovelSearchCondition condition, NovelPagingStrategy strategy, NovelCategory category);
+    List<? extends NovelQueryProjection> findNewNovels(NovelPagingCondition condition, NovelPagingStrategy strategy, NovelCategory category);
 }
