+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/NovelQueryRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.iucyh.novelservice.novel.repository.query;
 
 import com.iucyh.novelservice.novel.enumtype.NovelCategory;
-import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
-import com.iucyh.novelservice.novel.repository.query.condition.NovelSearchCondition;
+import com.iucyh.novelservice.novel.repository.query.projection.NovelQueryProjection;
+import com.iucyh.novelservice.novel.repository.query.condition.NovelPagingCondition;
 import com.iucyh.novelservice.novel.repository.query.paging.NovelPagingStrategy;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -21,7 +21,7 @@ public class NovelQueryRepositoryImpl implements NovelQueryRepository {
     private final JPAQueryFactory queryFactory;
 
     @Override
-    public List<? extends NovelQueryDto> findNovels(NovelSearchCondition condition, NovelPagingStrategy strategy, NovelCategory category) {
+    public List<? extends NovelQueryProjection> findNovels(NovelPagingCondition condition, NovelPagingStrategy strategy, NovelCategory category) {
         return strategy
                 .createQuery(queryFactory, condition.cursor())
                 .where(
@@ -33,7 +33,7 @@ public class NovelQueryRepositoryImpl implements NovelQueryRepository {
     }
 
     @Override
-    public List<? extends NovelQueryDto> findNewNovels(NovelSearchCondition condition, NovelPagingStrategy strategy, NovelCategory category) {
+    public List<? extends NovelQueryProjection> findNewNovels(NovelPagingCondition condition, NovelPagingStrategy strategy, NovelCategory category) {
         LocalDateTime thisMonth = getThisMonth();
         return strategy
                 .createQuery(queryFactory, condition.cursor())
