--- novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/dto/NovelQueryDto.java
@@ -1,8 +1,0 @@
-package com.iucyh.novelservice.novel.repository.query.dto;
-
-import com.iucyh.novelservice.novel.domain.Novel;
-
-public interface NovelQueryDto {
-
-    Novel getNovel();
-}
