+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/NovelPagingStrategy.java
@@ -1,6 +1,6 @@
 package com.iucyh.novelservice.novel.repository.query.paging;
 
-import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
+import com.iucyh.novelservice.novel.repository.query.projection.NovelQueryProjection;
 import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
 import com.querydsl.jpa.impl.JPAQuery;
@@ -8,7 +8,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 
 public interface NovelPagingStrategy {
 
-    JPAQuery<? extends NovelQueryDto> createQuery(JPAQueryFactory queryFactory, NovelCursor cursor);
-    NovelCursor createCursor(NovelQueryDto lastResult);
+    JPAQuery<? extends NovelQueryProjection> createQuery(JPAQueryFactory queryFactory, NovelCursor cursor);
+    NovelCursor createCursor(NovelQueryProjection lastResult);
     NovelSortType getSupportedSortType();
 }
