+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/paging/NovelPopularPagingStrategy.java
@@ -1,7 +1,7 @@
 package com.iucyh.novelservice.novel.repository.query.paging;
 
 import com.iucyh.novelservice.novel.domain.Novel;
-import com.iucyh.novelservice.novel.repository.query.dto.NovelQueryDto;
+import com.iucyh.novelservice.novel.repository.query.projection.NovelQueryProjection;
 import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelCursor;
 import com.iucyh.novelservice.novel.repository.query.paging.cursor.NovelPopularCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
@@ -46,7 +46,7 @@ public class NovelPopularPagingStrategy extends AbstractNovelPagingStrategy {
     }
 
     @Override
-    public NovelCursor createCursor(NovelQueryDto lastResult) {
+    public NovelCursor createCursor(NovelQueryProjection lastResult) {
         Novel lastNovel = lastResult.getNovel();
         return new NovelPopularCursor(lastNovel.getId(), lastNovel.getPeriodViewCount(), lastNovel.getTotalViewCount());
     }
