+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/projection/NovelQueryProjection.java
@@ -1,0 +1,8 @@
+package com.iucyh.novelservice.novel.repository.query.projection;
+
+import com.iucyh.novelservice.novel.domain.Novel;
+
+public interface NovelQueryProjection {
+
+    Novel getNovel();
+}
