+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/projection/NovelSimpleQueryProjection.java
@@ -1,16 +1,16 @@
-package com.iucyh.novelservice.novel.repository.query.dto;
+package com.iucyh.novelservice.novel.repository.query.projection;
 
 import com.iucyh.novelservice.novel.domain.Novel;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 
 @Getter
-public class NovelSimpleQueryDto implements NovelQueryDto {
+public class NovelSimpleQueryProjection implements NovelQueryProjection {
 
     private final Novel novel;
 
     @QueryProjection
-    public NovelSimpleQueryDto(Novel novel) {
+    public NovelSimpleQueryProjection(Novel novel) {
         this.novel = novel;
     }
 }
